requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mongoengine==0.18.2
pymongo==3.9.0
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
Werkzeug==0.16.0
WTForms==2.2.1
